% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_js.R
\name{create_js_file}
\alias{create_js_file}
\title{Title Create template JS file.}
\usage{
create_js_file(filepath = "./", overwrite = F)
}
\arguments{
\item{filepath}{Where to create base js file.}

\item{overwrite}{Should we overwrite any file named 'my_js.js' in the filepath?}
}
\value{
CSS file created in filepath.
}
\description{
Title Create template JS file.
}
\examples{
\dontrun
}
