% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_pages.R
\name{create_shiny_pages}
\alias{create_shiny_pages}
\title{Title Create the shiny scripts (R scripts) for shiny template}
\usage{
create_shiny_pages(filepath = "./", overwrite = F)
}
\arguments{
\item{filepath}{Where to create the new shiny template.}

\item{overwrite}{Should we overwrite any files in the filepath?}
}
\value{
Shiny R files created.
}
\description{
Title Create the shiny scripts (R scripts) for shiny template
}
\examples{
\dontrun
}
