% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_shiny_plate.R
\name{new_shiny_plate}
\alias{new_shiny_plate}
\title{Title Create a shiny template with separate CSS and JS files.}
\usage{
new_shiny_plate(filepath = NULL, new_project = T, overwrite = T)
}
\arguments{
\item{filepath}{Where to create the new shiny template.}

\item{overwrite}{Should we overwrite any files in the filepath?}
}
\value{
Shiny template with CSS and JS embedded.
}
\description{
Title Create a shiny template with separate CSS and JS files.
}
\examples{
\dontrun
}
